Consolidate IndexedDB admission transactions
