+++ packages/shared-web/browser/al-runtime/browser-al-runtime-cleanup.ts
@@ -1,11 +1,14 @@
-import { decodeALAdmissionStoredValue } from '@shared/alm/al-admission-backend.ts';
-import { ALAdmissionCorruptionError, decodeALAdmissionValue } from '@shared/alm/al-admission-decoder.ts';
-import { decodeALAdmissionNumber } from '@shared/alm/al-admission-value-validation.ts';
 import { isIndexedDbALRuntimeStoreSupported } from '@shared/alm/al-runtime-stores.ts';
 import { ALAdmissionBackendConflictError } from '@shared/alm/ALAdmissionBackendConflictError.ts';
-import { AL_ADMISSION_REVISION_KEY } from '@shared/alm/indexed-db-admission-backend.ts';
+import {
+    AL_ADMISSION_REVISION_KEY,
+    computeIndexedDbAdmissionRevisionWrite,
+    listIndexedDbAdmissionSnapshot,
+    readIndexedDbAdmissionKeySnapshot,
+    writeIndexedDbAdmissionMutations,
+    type IndexedDbAdmissionMutation
+} from '@shared/alm/indexed-db-admission-storage.ts';
 import { openIndexedDbWithStore } from '@shared/persistence/openIndexedDb.ts';
-import { NEVER_EXPIRE_AT_TIMESTAMP } from '@shared/persistence/PersistenceProvider.ts';
 import { tryRunInIntervals } from '@shared/resilience/TryWith.ts';
 
 import {
@@ -19,17 +22,18 @@ export const BROWSER_AL_RUNTIME_EXPIRY_EVICTION_INTERVAL_MS = 60_000;
 
 interface BrowserALRuntimeCleanupRead {
     readonly revision: number;
-    readonly rows: readonly Readonly<{ key: string; value: unknown; }>[];
+    readonly rows: readonly Readonly<{ key: string; expireAtTimestamp: number | null; }>[];
 }
 
 interface BrowserALRuntimeCleanupComputed {
-    readonly deleteKeys: readonly string[];
+    readonly mutations: readonly IndexedDbAdmissionMutation[];
     readonly revisionWrite: Readonly<{
         key: typeof AL_ADMISSION_REVISION_KEY;
         value: number;
         expireAtTimestamp: number;
     }>;
     readonly scanned: number;
+    readonly deleted: number;
 }
 
 type BrowserALRuntimeDeletionPolicy =
@@ -137,10 +141,11 @@ async function deleteBrowserALRuntimeEntriesMatching(
     );
 
     try {
-        const read = await readBrowserALRuntimeCleanup(db, keyPrefixes);
+        const read = await readBrowserALRuntimeCleanup(db, keyPrefixes, options.deletionPolicy);
         const computed = computeBrowserALRuntimeCleanup(read, options.deletionPolicy);
+        validateBrowserALRuntimeCleanup(read, options.deletionPolicy, computed);
         await writeBrowserALRuntimeCleanup(db, read.revision, computed);
-        return toBrowserALRuntimeCleanupResult(keyPrefixes, computed.scanned, computed.deleteKeys.length);
+        return toBrowserALRuntimeCleanupResult(keyPrefixes, computed.scanned, computed.deleted);
     }
     finally {
         db.close();
@@ -149,51 +154,30 @@ async function deleteBrowserALRuntimeEntriesMatching(
 
 async function readBrowserALRuntimeCleanup(
     db: IDBDatabase,
-    keyPrefixes: readonly string[]
+    keyPrefixes: readonly string[],
+    policy: BrowserALRuntimeDeletionPolicy
 ): Promise<BrowserALRuntimeCleanupRead> {
-    const transaction = db.transaction(BROWSER_AL_RUNTIME_STORE_NAME, 'readonly');
-    const store = transaction.objectStore(BROWSER_AL_RUNTIME_STORE_NAME);
-    const rowsPromise = readBrowserALRuntimeCleanupRows(store, keyPrefixes);
-    const revisionPromise = requestToPromise<unknown>(store.get(AL_ADMISSION_REVISION_KEY));
-    const completed = transactionDone(transaction);
-    const [rows, revisionValue] = await Promise.all([rowsPromise, revisionPromise]);
-    await completed;
-    return { rows, revision: decodeBrowserALRuntimeRevision(revisionValue) };
-}
-
-function readBrowserALRuntimeCleanupRows(
-    store: IDBObjectStore,
-    keyPrefixes: readonly string[]
-): Promise<readonly Readonly<{ key: string; value: unknown; }>[]> {
-    return new Promise((resolve, reject) => {
-        const rows: Readonly<{ key: string; value: unknown; }>[] = [];
-        const request = store.openCursor();
-        request.onerror = () => reject(request.error ?? new Error('Browser AL runtime cleanup cursor failed'));
-        request.onsuccess = () => {
-            const cursor = request.result;
-            if (cursor === null) {
-                resolve(rows);
-                return;
-            }
-            const key = cursor.primaryKey;
-            if (typeof key !== 'string') {
-                reject(new TypeError('Browser AL runtime cleanup row key must be a string'));
-                return;
-            }
-            if (matchesAnyBrowserALRuntimePrefix(key, keyPrefixes)) {
-                rows.push({ key, value: cursor.value });
-            }
-            cursor.continue();
+    if (policy.kind === 'all') {
+        const snapshot = await readIndexedDbAdmissionKeySnapshot(
+            db,
+            BROWSER_AL_RUNTIME_STORE_NAME,
+            keyPrefixes
+        );
+        return {
+            revision: snapshot.revision,
+            rows: snapshot.keys.map((key) => ({ key, expireAtTimestamp: null }))
         };
-    });
-}
-
-function decodeBrowserALRuntimeRevision(value: unknown): number {
-    if (value === undefined) {
-        return 0;
     }
-    const stored = decodeALAdmissionValue(value, AL_ADMISSION_REVISION_KEY, decodeALAdmissionStoredValue);
-    return decodeALAdmissionValue(stored.value, AL_ADMISSION_REVISION_KEY, decodeALAdmissionNumber);
+    const snapshot = await listIndexedDbAdmissionSnapshot(db, BROWSER_AL_RUNTIME_STORE_NAME, '');
+    return {
+        revision: snapshot.revision,
+        rows: snapshot.stored
+            .filter((stored) =>
+                stored.key !== AL_ADMISSION_REVISION_KEY &&
+                matchesAnyBrowserALRuntimePrefix(stored.key, keyPrefixes)
+            )
+            .map((stored) => ({ key: stored.key, expireAtTimestamp: stored.expireAtTimestamp }))
+    };
 }
 
 function computeBrowserALRuntimeCleanup(
@@ -202,19 +186,43 @@ function computeBrowserALRuntimeCleanup(
 ): BrowserALRuntimeCleanupComputed {
     const deleteKeys: string[] = [];
     for (const row of read.rows) {
-        if (shouldDeleteBrowserALRuntimeEntry(row.key, row.value, deletionPolicy)) {
+        if (
+            deletionPolicy.kind === 'all' ||
+            (row.expireAtTimestamp !== null && row.expireAtTimestamp <= deletionPolicy.nowMs)
+        ) {
             deleteKeys.push(row.key);
         }
     }
     return {
-        deleteKeys,
-        revisionWrite: {
-            key: AL_ADMISSION_REVISION_KEY,
-            value: read.revision + 1,
-            expireAtTimestamp: NEVER_EXPIRE_AT_TIMESTAMP
-        },
-        scanned: read.rows.length
+        mutations: deleteKeys.map((key) => ({ kind: 'remove', key })),
+        revisionWrite: computeIndexedDbAdmissionRevisionWrite(read.revision),
+        scanned: read.rows.length,
+        deleted: deleteKeys.length
     };
+}
+
+function validateBrowserALRuntimeCleanup(
+    read: BrowserALRuntimeCleanupRead,
+    deletionPolicy: BrowserALRuntimeDeletionPolicy,
+    computed: BrowserALRuntimeCleanupComputed
+): void {
+    const expected = computeBrowserALRuntimeCleanup(read, deletionPolicy);
+    if (
+        computed.scanned !== expected.scanned ||
+        computed.deleted !== expected.deleted ||
+        computed.revisionWrite.key !== expected.revisionWrite.key ||
+        computed.revisionWrite.value !== expected.revisionWrite.value ||
+        computed.revisionWrite.expireAtTimestamp !== expected.revisionWrite.expireAtTimestamp ||
+        computed.mutations.length !== expected.mutations.length ||
+        computed.mutations.some((mutation, index) => {
+            const expectedMutation = expected.mutations[index];
+            return mutation.kind !== 'remove' ||
+                expectedMutation?.kind !== 'remove' ||
+                mutation.key !== expectedMutation.key;
+        })
+    ) {
+        throw new TypeError('Browser AL runtime cleanup differs from its canonical computation');
+    }
 }
 
 async function writeBrowserALRuntimeCleanup(
@@ -222,62 +230,19 @@ async function writeBrowserALRuntimeCleanup(
     expectedRevision: number,
     computed: BrowserALRuntimeCleanupComputed
 ): Promise<void> {
-    if (computed.deleteKeys.length === 0) {
+    if (computed.mutations.length === 0) {
         return;
     }
-    const committed = await new Promise<boolean>((resolve, reject) => {
-        const transaction = db.transaction(BROWSER_AL_RUNTIME_STORE_NAME, 'readwrite');
-        const store = transaction.objectStore(BROWSER_AL_RUNTIME_STORE_NAME);
-        const revisionRequest = store.get(AL_ADMISSION_REVISION_KEY);
-        let conflict = false;
-
-        transaction.oncomplete = () => resolve(true);
-        transaction.onabort = () => {
-            if (conflict) {
-                resolve(false);
-                return;
-            }
-            reject(transaction.error ?? new Error('Browser AL runtime cleanup aborted'));
-        };
-        transaction.onerror = () => {
-            if (!conflict) {
-                reject(transaction.error ?? new Error('Browser AL runtime cleanup failed'));
-            }
-        };
-        revisionRequest.onerror = () =>
-            reject(revisionRequest.error ?? new Error('Browser AL runtime cleanup revision read failed'));
-        revisionRequest.onsuccess = () => {
-            const storedRevision = revisionRequest.result as { readonly value?: unknown; } | undefined;
-            const actualRevision = storedRevision?.value ?? 0;
-            if (actualRevision !== expectedRevision) {
-                conflict = true;
-                transaction.abort();
-                return;
-            }
-            for (const key of computed.deleteKeys) {
-                store.delete(key);
-            }
-            store.put(computed.revisionWrite);
-        };
+    const committed = await writeIndexedDbAdmissionMutations({
+        db,
+        storeName: BROWSER_AL_RUNTIME_STORE_NAME,
+        expectedRevision,
+        mutations: computed.mutations,
+        revisionWrite: computed.revisionWrite
     });
     if (!committed) {
         throw new ALAdmissionBackendConflictError('Browser AL runtime cleanup conflicted');
     }
-}
-
-function requestToPromise<T>(request: IDBRequest<T>): Promise<T> {
-    return new Promise<T>((resolve, reject) => {
-        request.onsuccess = () => resolve(request.result);
-        request.onerror = () => reject(request.error ?? new Error('Browser AL runtime cleanup request failed'));
-    });
-}
-
-function transactionDone(transaction: IDBTransaction): Promise<void> {
-    return new Promise<void>((resolve, reject) => {
-        transaction.oncomplete = () => resolve();
-        transaction.onabort = () => reject(transaction.error ?? new Error('Browser AL runtime cleanup read aborted'));
-        transaction.onerror = () => reject(transaction.error ?? new Error('Browser AL runtime cleanup read failed'));
-    });
 }
 
 function toBrowserALRuntimeCleanupResult(
@@ -304,30 +269,4 @@ function matchesAnyBrowserALRuntimePrefix(
         }
     }
     return false;
-}
-
-function shouldDeleteBrowserALRuntimeEntry(
-    key: string,
-    value: unknown,
-    policy: BrowserALRuntimeDeletionPolicy
-): boolean {
-    if (policy.kind === 'all') {
-        return true;
-    }
-    if (value === null || typeof value !== 'object') {
-        throw new ALAdmissionCorruptionError(key, new TypeError('Browser AL runtime cleanup row must be an object'));
-    }
-    const expireAtTimestamp = (value as { readonly expireAtTimestamp?: unknown; }).expireAtTimestamp;
-    if (
-        typeof expireAtTimestamp !== 'number' ||
-        !Number.isSafeInteger(expireAtTimestamp) ||
-        expireAtTimestamp < 0 ||
-        Object.is(expireAtTimestamp, -0)
-    ) {
-        throw new ALAdmissionCorruptionError(
-            key,
-            new TypeError('Browser AL runtime cleanup expiry must be a non-negative safe integer')
-        );
-    }
-    return expireAtTimestamp <= policy.nowMs;
 }
